+++ thoth-api/migrations/v0.11.14/up.sql
@@ -1,0 +1,4 @@
+ALTER TYPE location_platform ADD VALUE 'Google Books';
+ALTER TYPE location_platform ADD VALUE 'Internet Archive';
+ALTER TYPE location_platform ADD VALUE 'ScienceOpen';
+ALTER TYPE location_platform ADD VALUE 'SciELO';